+++ tests/test_image.py
@@ -135,12 +135,12 @@ def test_get_exif_tags():
     src_file = os.path.join(CURRENT_DIR, 'sample', 'pictures', 'dir1', 'test1',
                             test_image)
     data = get_exif_data(src_file)
-    simple = get_exif_tags(data)
+    simple = get_exif_tags(data, datetime_format='%d/%m/%Y')
     assert simple['fstop'] == 3.9
     assert simple['focal'] == 12.0
     assert simple['iso'] == 50
     assert simple['Make'] == 'NIKON'
-    assert simple['datetime'] == 'Sunday, 22. January 2006'
+    assert simple['datetime'] == '22/01/2006'
     if PIL.PILLOW_VERSION == '3.0.0':
         assert simple['exposure'] == 0.00100603
     else: